fix: Snooze only current items when `focus`ing
In order to do this, I needed to clobber existing `when` properties
when calling `snooze()` on a list of todos. Existing `when`s were
passed through previously because `snooze()` might be called with
already snoozed todos.

With the clobber paradigm, it makes more sense to only snooze current
items when calling `focus`. This means items that are past their
`snooze` date are still snoozed to the new `when` date (currently
always the 'Someday' value of the year 9999).